relocated eventviewtoggle and removed dateOrder array from eventview
