+++ client/components/events/EventCalendarView/EventCalendarView.tsx
@@ -4,10 +4,10 @@ import styled from 'styled-components'
 import moment from 'moment-timezone'
 import { colours, device, fontWeight, desktopFontSize, mobileFontSize } from '../../../styles'
 import { EventCalendarCard } from './EventCalendarCard'
-import { EVENT_MAP_KEY_FORMAT, FilteredEvents } from "../../../pages/events"
+import { EVENT_MAP_KEY_FORMAT } from "../../../pages/events"
 
 interface EventViewProps {
-  filteredEvents: FilteredEvents
+  events: Map<string, Array<Event>>
 }
 
 interface DaySignature {
@@ -55,12 +55,12 @@ const CalendarColumnHeader = styled.h3`
   margin-bottom: 24px;
 `
 
-export const EventCalendarView = ({ filteredEvents }: EventViewProps) => {
+export const EventCalendarView = ({ events }: EventViewProps) => {
 
   const getEventsOnDay = (dayNumber: number) => {
     const weekDay = moment().startOf('isoWeek').add(dayNumber, 'days');
     //ensures events correspond to current week period starting at the last monday and ending at next sunday are selected
-    return filteredEvents.filteredEventsDateMap[weekDay.format(EVENT_MAP_KEY_FORMAT)]
+    return events.get(weekDay.format(EVENT_MAP_KEY_FORMAT))
   }
 
   const generateDayColumns = () => (
